In this file, million playlist dataset and lyric dataset will be processed.
The file removes unnecessary details like 'track_uri' and 'artist_uri'. Tracks not in the lyrics dataset are excluded. Remaining tracks are then enhanced with keywords, mood, and genre. Also it changes the processed lyric dataset from a CSV format to a JSON file that can be used by ChromaDB.